+++ src/main/webapp/WEB-INF/views/board/comment/comment.jsp
@@ -12,12 +12,12 @@
 			<td>${DTO.boardCommentDate}</td>
 			<td>
 				<c:if test="${member.memberId eq DTO.memberId}">
-					<button class="btn btn-info update" data-boardcomment-num="${DTO.boardCommentNum}" data-bs-toggle="modal" data-bs-target="#contentsModal">수정</button>
+					<button class="btn btn-info" data-boardcomment-num="${DTO.boardCommentNum}" data-bs-toggle="modal" data-bs-target="#contentsModal">수정</button>
 				</c:if>
 			</td>
 			<td>
 				<c:if test="${member.memberId eq DTO.memberId}">
-					<button class="btn btn-danger delete" data-boardcomment-num="${DTO.boardCommentNum}">삭제</button>
+					<button class="btn btn-danger delete" id="del" data-boardcomment-num="${DTO.boardCommentNum}">삭제</button>
 				</c:if>
 			</td>
 		</tr>
