+++ src/main/webapp/WEB-INF/views/board/detail.jsp
@@ -86,7 +86,6 @@
 		</div>
 	</div>
 	<c:import url="../template/common_js.jsp"></c:import>
-	<script src="/resources/js/boardComment.js"></script>
-	<script type="text/javascript" src="../resources/js/confirm.js"></script>
+	<script src="/resources/js/board/comment.js"></script>
 </body>
 </html>
