+++ src/main/webapp/WEB-INF/views/lesson/add.jsp
@@ -44,6 +44,8 @@
 	<script src="../resources/js/fileManager.js"></script>
 	<script>
 		setMax(4);
+		setCount('${DTO.lessonImgDTOs.size()}');
+		setParam("imgs")
 		$('#lessonContents').summernote();
 	</script>
 	<c:import url="../template/common_js.jsp"></c:import>
