+++ src/main/webapp/WEB-INF/views/lesson/detail.jsp
@@ -110,7 +110,6 @@
 	</div>
 	
 	<c:import url="../template/common_js.jsp"></c:import>
-	<script src="/resources/js/lessonReview.js"></script>
-	<script type="text/javascript" src="../resources/js/confirm.js"></script>
+	<script src="/resources/js/lesson/review.js"></script>
 </body>
 </html>
