+++ src/main/webapp/WEB-INF/views/product/add.jsp
@@ -59,9 +59,11 @@
 			</div>
 	   </form>
 	</div>
-	<script src="../resources/js/fileManager.js"></script>
+	<script src="../resources/js/common/fileManager.js"></script>
 	<script>
 		setMax(4);
+		setCount('${DTO.productImgDTOs.size()}');
+		setParam("imgs")
 		$('#productDetail').summernote();
 	</script>
 	<c:import url="../template/common_js.jsp"></c:import>
