+++ src/main/webapp/resources/js/common/fileManager.js
