--- src/main/webapp/resources/js/confirm.js
@@ -1,11 +1,0 @@
-const frm = document.getElementById("frm");
-const d = document.getElementById("delete");
-
-d.addEventListener("click", function(){
-    let check = window.confirm("삭제하시겠습니까?");
-    if(check) {
-        frm.setAttribute("action", "./delete");
-        frm.setAttribute("method", "post");
-        frm.submit();
-    }
-})
