+++ src/test/java/com/team/cwl/board/comment/BoardCommentDAOTest.java
@@ -33,7 +33,7 @@ public class BoardCommentDAOTest extends MyTestCase {
 		
 			BoardCommentDTO boardCommentDTO = new BoardCommentDTO();
 			
-			boardCommentDTO.setBoardNum(31L);
+			boardCommentDTO.setBoardNum(21L);
 			boardCommentDTO.setMemberId("mhj");
 			boardCommentDTO.setBoardCommentContents("말랑복숭아"+i);
 			boardCommentDTO.setBoardCommentDate(date);
